+++ UnitTests/BraintreeDataCollectorTests/MockBTDataCollector.swift
@@ -6,7 +6,7 @@ class MockBTDataCollector: BTDataCollector {
     var cannedDeviceData: String?
     var cannedDataCollectorError: Error?
     
-    override func collectDeviceData(_ completion: @escaping (String?, Error?) -> Void) {
+    override func collectDeviceData(riskCorrelationID: String? = nil, completion: @escaping (String?, Error?) -> Void) {
         completion(cannedDeviceData, cannedDataCollectorError)
     }
 }